Improve the jersey data transfer module
- Make it undependent from spring data
- Make the pageable value provider abstract
- Add a PagingRequest pageable provider
- Add a spring-data jersey data transfer module containing a spring data pageable value provider
